% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmodelling-package.R
\docType{package}
\name{Rmodelling-package}
\alias{Rmodelling}
\alias{Rmodelling-package}
\title{Rmodelling: A new approach to modelling in R}
\description{
The package provides a set of functions for modelling.
}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
\keyword{internal}
